Powershell script to query DHCP lease
Script should be run in DC or on a machine that has powershell domain controller module imported. It will scan for existing DHCP servers in domain and then scan through the respective servers to populate leases utilized. It will through errors if DHCP server is unreachable or if Domain account does not have permission.